requirements.txt
flask==0.12
flask-restful==0.3.6
apscheduler==3.3.1
Flask-SQLAlchemy==2.2
Flask-Migrate==2.0.4
psycopg2==2.7.1
